+++ odoo/songs/install/data_all.py
@@ -115,6 +115,10 @@ def import_account_account(ctx):
     load_csv(ctx, 'data/install/account.csv',
              'account.account')
 
+def import_email_template(ctx):
+    """ Importing email template from csv """
+    load_csv(ctx, 'data/install/email_template.csv', 'mail.template')
+
 
 @anthem.log
 def main(ctx):
@@ -131,3 +135,4 @@ def main(ctx):
     delete_account(ctx)
     import_account_account(ctx)
     create_date_range(ctx)
+    import_email_template(ctx)