+++ HISTORY.rst
@@ -129,6 +129,7 @@ latest (unreleased)
 * Install module sale_order_lot_selection
 * Install module note
 * Update settings for accounting, logistics, manufacturing and sales
+* Add the email template invoicing data
 
 **Bugfixes**